requirements.txt
Flask==2.3.3
flask-mysqldb==1.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Jinja2==3.1.3
Werkzeug==2.3.7
PyMySQL